+++ CHANGELOG.md
@@ -34,6 +34,10 @@
 - Added `UpdateAffiliation` and `AffiliationById` queries [#203]
 
 ## Updated
+- Removed the best practice checkbox group of tags from the add/edit Section pages, and added them to the add/edit Question pages [#274]
+- Updated the `TransitionButton` to use `startTransaction` so components using it can better control the loading gif and progress bar. Updated some of the pages to utilize this new `startTransaction` [#274]
+- Updated shared `QuestionAdd` to include a `showTags` boolean so that we can exclude it for custom pages for now [#274]
+- Made improvements on progress bar and transition buttons on some pages in the template and project flow [#274]
 - Updated `layout.tsx` to use the CDL logo from the CDN. Added new `favicon.ico` as a fallback, and added a new `CDN_ENDPOINT` to match the parameter set in Cloud Formation template [#328]
 - Updated `login` and `signup` pages with `document.title`, and added `Pageheader` to `home` page [#328]
 - Updated `signup` page to use the new `PasswordRequirementsList` component, and add some new unit tests [#242]